Refactor: Organize JS files based on functionalities
