+++ index.html
@@ -108,9 +108,10 @@
         <!-- TODO: Pagination -->
       </section>
     </main>
-    <script src="./js/script.js" type="module"></script>
-    <script src="./js/toggleTheme.js" type="module"></script>
+    <script src="./js/main.js" type="module"></script>
+    <script src="./js/ui.js" type="module"></script>
     <script src="./js/lib/firebase.js" type="module"></script>
-    <script src="./js/fetchData.js" type="module"></script>
+    <script src="./js/filter.js" type="module"></script>
+    <script src="./js/toggleTheme.js" type="module"></script>
   </body>
 </html>
