+++ js/utils/toast.js
@@ -1,0 +1,6 @@
+export const showToast = (type) => {
+  if (type === "warning") {
+  } else if (type === "error") {
+  } else if (type === "success") {
+  } else return;
+};
